+++ src/components/main-view/main-view.jsx
@@ -5,8 +5,8 @@ import MoviesList from '../movies-list/movies-list';
 import Footer from '../footer/footer';
 import GenresList from '../genres-list/genres-list';
 import ShowMore from '../show-more/show-more';
-import {PROMO_PROPS, FILMS_PROPS} from '../../utils';
-import {getSortedFilmsByGenre} from '../../logic';
+import {PROMO_PROPS, FILMS_PROPS} from '../../utils/proptypes';
+import {getSortedFilmsByGenre} from '../../utils/helpers';
 import {fetchFilmsList, fetchPromoMovie} from '../../store/api-actions';
 import LoadingView from '../loading-view/loading-view';
 import UserBlock from '../user-block/user-block';
