+++ src/components/movies-list/movies-list.jsx
@@ -1,7 +1,7 @@
 import React, {useState} from 'react';
 import PropTypes from 'prop-types';
 import SmallMovieCard from '../small-movie-card/small-movie-card';
-import {FILMS_PROPS} from '../../utils';
+import {FILMS_PROPS} from '../../utils/proptypes';
 
 const MoviesList = ({films, showed}) => {
   const [activeMovie, setActiveMovie] = useState(-1);
