+++ src/components/private-route/private-route.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {Route, Redirect} from 'react-router-dom';
 import {connect} from 'react-redux';
-import {AuthorizationStatus} from '../../utils';
+import {AuthorizationStatus} from '../../utils/const';
 
 
 const PrivateRoute = ({render, path, exact, authorizationStatus}) => {
