+++ src/components/user-block/user-block.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import {Link} from 'react-router-dom';
-import {AuthorizationStatus} from '../../utils';
+import {AuthorizationStatus} from '../../utils/const';
 
 const AuthorizedUser = ({avatarUrl}) => {
   return (
