+++ src/utils/const.js
@@ -1,0 +1,17 @@
+export const RATING = {
+  "Bad": [0, 1, 2],
+  "Normal": [3, 4],
+  "Good": [5, 6, 7],
+  "Very Good": [8, 9],
+  "Awesome": [10],
+};
+export const RATING_STARS = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
+
+export const ALL_GENRES = `All genres`;
+
+export const FILMS_PER_STEP = 8;
+
+export const AuthorizationStatus = {
+  AUTH: `AUTH`,
+  NO_AUTH: `NO_AUTH`,
+};
